+++ trixnity-messenger/src/commonMain/kotlin/de/connect2x/trixnity/messenger/util/RelevantTimelineEvents.kt
@@ -3,10 +3,7 @@ package de.connect2x.trixnity.messenger.util
 import net.folivo.trixnity.core.model.events.MessageEventContent
 import net.folivo.trixnity.core.model.events.RoomEventContent
 import net.folivo.trixnity.core.model.events.m.RelatesTo
-import net.folivo.trixnity.core.model.events.m.room.CreateEventContent
 import net.folivo.trixnity.core.model.events.m.room.EncryptedMessageEventContent
-import net.folivo.trixnity.core.model.events.m.room.MemberEventContent
-import net.folivo.trixnity.core.model.events.m.room.NameEventContent
 import net.folivo.trixnity.core.model.events.m.room.RoomMessageEventContent
 
 interface RelevantTimelineEvents {
@@ -14,9 +11,8 @@ interface RelevantTimelineEvents {
         val isReplace =
             content is MessageEventContent && content.relatesTo is RelatesTo.Replace
         val isMessage = content is RoomMessageEventContent || content is EncryptedMessageEventContent
-        val isStateEvent = content is MemberEventContent || content is NameEventContent || content is CreateEventContent
 
-        return !isReplace && (isMessage || isStateEvent)
+        return !isReplace && (isMessage)
     }
 
     companion object : RelevantTimelineEvents