+++ CHANGELOG.md
@@ -19,6 +19,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+- Don't show placeholder, when lastRelevantEventId not known.
+- Don't mark state events as relevant.
 - Wizard and modal dialog buttons are always shown on smaller screens 
 
 ### Security